# download/MaestroAgent/maestro-personal/docs/openapi_schema.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LoginRequest": {
        "properties": {
          "password": {
            "default": "",
            "title": "Password",
            "type": "string"
          },
          "user_email": {
            "default": "",
            "title": "User Email",
            "type": "string"
          }
        },
        "title": "LoginRequest",
        "type": "object"
      },
      "LoginResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          },
          "user_email": {
            "title": "User Email",
            "type": "string"
          }
        },
        "required": [
          "token",
          "user_email",
          "message"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
...
      "PushTokenRequest": {
        "properties": {
          "push_token": {
            "title": "Push Token",
            "type": "string"
          }
        },
        "required": [
          "push_token"
        ],
        "title": "PushTokenRequest",
        "type": "object"
      },
      "RegisterRequest": {
        "properties": {
          "password": {
            "title": "Password",
            "type": "string"
          },
          "user_email": {
            "format": "email",
            "title": "User Email",
            "type": "string"
          }
        },
        "required": [
          "user_email",
          "password"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
      "RegisterResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          },
          "user_email": {
            "title": "User Email",
            "type": "string"
          }
        },
        "required": [
          "user_email",
          "message",
          "token"
        ],
        "title": "RegisterResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "HTTP API for Maestro Personal v1 \u2014 wraps the PersonalShell (Core via Python)",
    "title": "Maestro Personal API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "description": "Login \u2014 returns a bearer token.\n\nP1 fix: passwordless email login removed. The login now requires\neither:\n1. The MAESTRO_PERSONAL_TOKEN env var (single-user local mode) \u2014\n   the caller must provide it as the password. No email-based login.\n2. A per-user token that was previously created via _create_user_token.\n   But tokens are never created without the setup password.\n\nIn dev mode (MAESTRO_PERSONAL_ENV not 'production'):\n- Bootstrap token works with password=AUTH_TOKEN (backward compat for tests)\n- Email-only login is REJECTED\n\nIn production mode:\n- Only per-user tokens work (no bootstrap)\n- Login requires password validation against user store (future)\n\nThis closes the P0-2 passwordless login vulnerability.",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/push-token": {
      "post": {
        "description": "Store the user's Expo push token for sending notifications.\n\nIssue 6: Called by the mobile app on login to register the device\nfor push notifications. The notification_scheduler uses these tokens\nto send stale-commitment alerts.",
        "operationId": "register_push_token_api_auth_push_token_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Push Token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "description": "Register a new account with email + password.\n\nPhase 2: Real account lifecycle. Creates a user account with\na hashed password (PBKDF2-SHA256, 100k iterations). Returns\na bearer token immediately after registration.\n\nRate limited: 3 registrations per hour per IP.",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/revoke": {
      "post": {
        "description": "Revoke the current token (P1-4 fix).\n\nThe caller's bearer token (from the Authorization header) is revoked.\nAfter this call, the token can no longer be used for authentication.\nThe user must log in again to get a new token.\n\nThis is the standard 'logout' endpoint \u2014 it ensures that even if the\ntoken is intercepted, it becomes useless after revocation.",
        "operationId": "revoke_token_api_auth_revoke_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/rotate": {
      "post": {
        "description": "Rotate the current token (P1-4 fix).\n\nIssues a new token and revokes ALL old tokens for the user. This is\nthe standard token rotation flow \u2014 call this periodically to limit\nthe window of opportunity for a compromised token.\n\nReturns the new token. The old token(s) are immediately invalid.",
        "operationId": "rotate_token_api_auth_rotate_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rotate Token",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}